I THINK I FIXED IT

+++ main.sh
@@ -2,13 +2,15 @@
 echo "Minifying lua"
 if [ -n "$GITHUB_WORKSPACE" ]
     cd "$GITHUB_WORKSPACE"
-    pwd
-    for file in (find $argv)
+    echo $argv[1]
+    for file in (echo $argv[1] | xargs find)
         echo "$file"
         luamin -f "$file" > "$file".tmp
         if luamin -f "$file" > "$file".tmp
             echo "Lua valid, renaming $file.tmp -> $file"
-            mv "$file.tmp" "$file"
+            if not set -q DRY_RUN then
+                mv "$file.tmp" "$file"
+            end
         else
             echo "FILE IS NOT LUA"
             cat "$file.tmp"
